+++ src/script/bindings/mod.rs
@@ -9,6 +9,7 @@ pub mod memcached;
 pub mod ntp;
 pub mod ping;
 pub mod redis;
+pub mod rtsp;
 pub mod tcp;
 pub mod tls;
 pub mod whois;